1. Verification approach is refactored, made common for all tests (bfs, bfs_spmd, sssp); common.hpp removed, instead added bfs.hpp and sssp.hpp. 2. Tabs are replaced with double spaces.
